Compton Compositor tweak added
*** Enabled for Linux Mint Xfce and Linux Lite only - needs testing on
other distros. Will be enabled for other supported distros once tested
***
- Installs compton and provides a functional/tested/optimized
configuration out of the box.
- It disables the default compositor during installation (if enabled)
to prevent display compositing conflicts.
- Compton configuration is installed and available for all user
accounts on install.
- Compton Autostart is installed and available for all user accounts.
Each user account can enable or disable compton compositor individually.
- Elevation is only required for install/uninstall. It uses pkexec so
that it can be installed or removed from any account with proper
elevation.
- When user enables compton it sets itself to autostart. When disabled,
it removes itself from the autostart to prevent conflicts with other
display compositor the user may enable.
- If compton was installed prior to UnlockMe, a configuration file will
be set for the user in those cases if it doesn't exists.
- GUI displays Compton status Enabled/Disabled and hints whether it is
set for Autostart or not.